tests/wpt/tests/css/css-properties-values-api/registered-property-crosstalk.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.css-houdini.org/css-properties-values-api-1" />
<link rel="help" href="https://bugs.chromium.org/p/chromium/issues/detail?id=1238686" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

  @property --x {
    syntax: "<number>";
    inherits: true;
    initial-value: 0;
  }

  #a {
    --y: 0;
  }

  #b {
    --z: 0;
  }

  #c {
    --x: 42;
  }

</style>

<div id=a>
  <div id=b>
    <div id=c>
    </div>
  </div>
</div>

<script>

test(function(){
  assert_equals(getComputedStyle(a).getPropertyValue('--x'), '0');
  assert_equals(getComputedStyle(b).getPropertyValue('--x'), '0');
  assert_equals(getComputedStyle(c).getPropertyValue('--x'), '42');
}, 'Only #c should be affected by --x:42');

</script>
