LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/registered-property-crosstalk-expected.txt


FAIL Only #c should be affected by --x:42 assert_equals: expected "0" but got ""
